DTOs and OrderService b-gun

+++ src/main/java/eu/mrndesign/matned/metalserwisproductionrest/model/audit/AuditInterface.java
@@ -23,9 +23,9 @@ public interface AuditInterface {
         return new AuditDTO.AuditBuilder()
                 .id(entity.getId())
                 .version(entity.getVersion())
-                .createdById(entity.getCreatedBy().map(AbstractPersistable::getId).orElse(0L))
+                .createdById((Long) entity.getCreatedBy().map(AbstractPersistable::getId).orElse(0L))
                 .createdDate(entity.getCreatedDate().map(x->x.format(DATE_TIME_FORMATTER)).orElse(LocalDateTime.now().format(DATE_TIME_FORMATTER)))
-                .lastModifiedById(entity.getLastModifiedBy().map(AbstractPersistable::getId).orElse(0L))
+                .lastModifiedById((Long) entity.getLastModifiedBy().map(AbstractPersistable::getId).orElse(0L))
                 .lastModifiedDate(entity.getLastModifiedDate().map(x->x.format(DATE_TIME_FORMATTER)).orElse(LocalDateTime.now().format(DATE_TIME_FORMATTER)))
                 .build();
     }
